Organized VC Contents and Added Documentation
