+++ Scoop/SearchRidesViewController.swift
@@ -10,7 +10,8 @@ import UIKit
 
 class SearchRidesViewController: UIViewController {
     
-    // MARK: Views
+    // MARK: - Views
+    
     private let arrivalLabel = UILabel()
     private let arrivalTextField = ImageTextField()
     private let calendarIconImageView = UIImageView()
@@ -22,11 +23,12 @@ class SearchRidesViewController: UIViewController {
     private let findTripsButton = UIButton()
     private let stackView = UIStackView()
     
-    // MARK: Data
+    // MARK: - Data
     private var arrivalPlace: GMSPlace?
     private var departurePlace: GMSPlace?
     private var tripDate: String = ""
     
+    // MARK: - Lifecycle Functions
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .white
@@ -38,6 +40,8 @@ class SearchRidesViewController: UIViewController {
         setupButton()
         setupLabels()
     }
+    
+    // MARK: - Setup View Functions
     
     private func setupHeader() {
         let dottedLineMultiplier = 0.52
@@ -201,6 +205,8 @@ class SearchRidesViewController: UIViewController {
             make.width.equalTo(99)
         }
     }
+    
+    // MARK: - Helper Functions
     
     @objc private func presentDepartureSearch() {
         let depatureVC = DepartureSearchViewController()
